refactor: extract embedded `mod tests` for 3 more god-files (batch 2)
Continues the cross-subsystem "extract embedded mod tests" sweep (after dml #871).
Each file's single `#[cfg(test)] mod tests { ... }` block moved to a sibling
`<stem>_tests.rs` via `#[cfg(test)] #[path] mod tests;`. Zero behavioral change
(child module preserves `use super::*;` private access); identical test suites.

- storage/document/service.rs: 4,396 -> 2,741 (tests=1,655)
- network/postgres/protocol.rs: 6,440 -> 5,342 (tests=1,298)
- storage/engines/core/formats/columnar/text_storage.rs: 2,567 -> 1,818 (tests=749)

~3,700 lines moved out of the root-crate compile unit. Verified: cargo check --tests
-p proximadb clean (exit 0); rustfmt 1.95.0 clean.
